requirements.txt
acron==0.17.2
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alabaster==0.7.12
alembic==1.7.5
amdsmi @ file:///opt/rocm-6.4.1/share/amd_smi
anyio==4.11.0
appdirs==1.4.4
argcomplete==1.12.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
argparse-manpage==4.7
arrow==1.3.0
astroid==2.11.6
asttokens==2.0.5
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.4.0
audioread==3.1.0
auth-get-sso-cookie==2.3.0
Authlib==1.0.1
autopage==0.5.2
awkward==2.8.9
awkward_cpp==50
awscli==2.15.31
awscrt==0.20.5
babel==2.17.0
backcall==0.1.0
backoff==2.2.1
beautifulsoup4==4.10.0
bleach==6.2.0
bodhi-client==7.0.1
cached-property==1.5.2
cachetools==6.2.1
CCColUtils==1.5
certifi==2023.7.22
cffi==1.14.5
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
cliff==4.0.0
cmd2==2.4.2
colorama==0.4.6
comm==0.2.3
cramjam==2.11.0
cryptography==36.0.1
cupy-cuda12x==13.6.0
cycler==0.11.0
Cython==0.29.37
dbus-python==1.2.18
debtcollector==2.5.0
debugpy==1.8.17
decorator==4.4.2
defusedxml==0.7.1
dill==0.3.5.1
distlib==0.3.2
distro==1.5.0
docutils==0.16
dogpile.cache==1.1.8
durationpy==0.10
elasticsearch==7.17.4
elasticsearch6==6.8.2
exceptiongroup==1.3.0
executing==0.8.2
fastjsonschema==2.21.2
fastrlock==0.8.3
fedpkg==1.46
file-magic==0.4.0
filelock==3.12.2
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2023.1.0
git-review==2.3.1
gitdb==4.0.9
GitPython==3.1.45
google-auth==2.41.1
gpg==1.15.1
greenlet==1.1.2
grpcio==1.75.1
gssapi==1.6.9
h11==0.16.0
h5py==3.7.0
htcondor==23.0.28
httpcore==1.0.9
httplib2==0.20.3
httpx==0.27.2
huggingface-hub==0.16.4
idna==3.4
imagesize==1.2.0
importlib-metadata==6.7.0
ipykernel==6.30.1
ipython==8.5.0
iso8601==1.0.2
isoduration==20.11.0
isort==5.10.1
jedi==0.18.1
jeepney==0.8.0
jellyfish==1.0.0
Jinja2==3.1.6
jmespath==0.9.4
joblib==1.3.2
json5==0.12.1
jsonpatch==1.21
jsonpickle==3.0.2
jsonpointer==2.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kerberos==1.3.0
keyring==21.8.0
keystoneauth1==5.0.1
kiwisolver==1.4.4
koji==1.35.2
kubeflow-katib==0.18.0
kubeflow-training==1.9.0
kubernetes==34.1.0
lark==1.3.0
lazy-object-proxy==1.7.1
lazy_loader==0.4
ldap3==2.8.1
libcomps==0.1.18
librosa==0.10.1
llvmlite==0.41.1
lxml==4.6.5
M2Crypto==0.38.0
Mako==1.1.4.dev0
MarkupSafe==2.0.0
matplotlib==3.5.3
matplotlib-inline==0.1.5
mccabe==0.6.1
mercurial==6.1.4
mido==1.3.0
mistune==3.1.4
more-itertools==9.1.0
mpi4py==4.1.0
mpmath==1.3.0
msgpack==1.0.3
multidict==6.7.0
munch==2.5.0
music21==7.3.3
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
netaddr==0.10.1
netifaces==0.10.6
networkx==3.2.1
nltk==3.8.1
notebook==7.2.2
notebook_shim==0.2.4
numba==0.58.0
numpy==1.25.2
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
oauthlib==3.3.1
olefile==0.46
openidc-client==0.7.0
openstacksdk==0.101.0
os-client-config==2.1.0
os-service-types==1.7.0
osc-lib==2.6.2
oslo.config==9.0.0
oslo.context==5.0.0
oslo.i18n==5.1.0
oslo.log==5.0.0
oslo.serialization==5.0.0
oslo.utils==6.0.2
overrides==7.7.0
packaging==23.1
pandas==1.3.5
pandocfilters==1.5.1
parso==0.8.3
pbr==5.10.0
pcp==5.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==2.5.4
ply==3.11
pooch==1.8.2
prettytable==0.7.2
prometheus_client==0.23.1
prompt-toolkit==3.0.38
propcache==0.4.1
protobuf==4.25.8
psutil==7.1.0
ptyprocess==0.6.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.1
pycparser==2.20
pycurl==7.43.0.6
pydub==0.25.1
pyenchant==3.2.0
Pygments==2.7.4
PyGObject==3.40.1
pyinotify==0.9.6
pylint==2.13.7
PyMySQL==0.10.1
pyparsing==3.1.1
pyperclip==1.8.0
pyphonebook==2.2.0
PyQt5==5.15.9
PyQt5-sip==12.11.1
pyroute2==0.5.14
PySocks==1.7.1
python-bugzilla==3.2.0
python-cinderclient==9.1.1
python-dateutil==2.8.2
python-gitlab==3.9.0
python-json-logger==4.0.0
python-keystoneclient==5.0.1
python-ldap==3.4.3
python-linux-procfs==0.7.3
python-magnumclient==4.0.0
python-neutronclient==8.1.0
python-novaclient==18.1.0
python-openstackclient===6.0.1-1.1
pytz==2023.3
pyudev==0.22.0
PyYAML==6.0.1
pyzmq==27.1.0
rarfile==4.1
referencing==0.36.2
regex==2023.8.8
requests==2.31.0
requests-gssapi==1.2.3
requests-kerberos==0.12.0
requests-oauthlib==2.0.0
requests-toolbelt==0.9.1
requestsexceptions==1.4.0
retrying==1.4.2
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
ROOT==6.36.4
rpds-py==0.27.1
rpkg==1.68
rpm==4.16.1.3
rsa==4.9.1
ruamel.yaml==0.16.6
ruamel.yaml.clib==0.2.7
s3cmd==2.4.0
sacremoses==0.0.53
samplings==0.1.7
scikit-learn==1.6.1
scipy==1.9.3
seaborn==0.13.2
SecretStorage==3.3.1
selinux==3.6
Send2Trash==1.8.3
sepolicy==3.6
setools==4.4.4
shtab==1.5.8
simplejson==3.17.6
six==1.16.0
smmap==5.0.0
sniffio==1.3.1
snowballstemmer==1.9.0
soundfile==0.13.1
soupsieve==2.4.1
soxr==1.0.0
Sphinx==3.4.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
SQLAlchemy==1.4.45
stack-data==0.5.0
stevedore==4.0.2
suds==1.1.2
sympy==1.14.0
systemd-python==234
tabulate==0.8.10
templated-dictionary==1.5
termcolor==1.1.0
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tldr==3.3.0
tokenizers==0.12.1
tomli==2.0.1
torch==2.8.0
torch-geometric==2.6.1
tornado==6.4.1
tqdm==4.66.1
-e git+ssh://git@github.com/cms-patatrack/TICL-GNN-Trackster-Linking.git@b1370d748d60e20008373d9880574bbee3035159#egg=tracksterLinker&subdirectory=tracksterLinker
traitlets==5.14.3
transformers==4.18.0
triton==3.4.0
typed-ast==1.5.4
types-python-dateutil==2.9.0.20251008
typing_extensions==4.15.0
Unidecode==1.3.6
uproot==5.6.6
uri-template==1.3.0
urllib3==2.0.4
virtualenv==20.21.1
wcwidth==0.2.5
webcolors==1.13
webencodings==0.5.1
websocket-client==1.9.0
wrapt==1.14.1
xlrd==2.0.1
xxhash==3.6.0
yarl==1.22.0
zipp==3.15.0
